挖矿
1.导入挖矿素材
2.完成A1，B1矿
3.完成popmenu
4.完成技能树镜头缩放效果

+++ Assets/Scripts/Environmnent/DestroyableTile.cs
@@ -5,7 +5,7 @@ using UnityEngine.Tilemaps;
 
 public class DestroyableTile : MonoBehaviour
 {
-    public int HP = 10;
+    public float HP = 10;
     public GameObject[] drops;
     [Header("Number of drops")]
     public int min = 1;
@@ -19,7 +19,7 @@ public class DestroyableTile : MonoBehaviour
         map = transform.parent.GetComponentInParent<Tilemap>();
     }
 
-    public bool SetHP(int value)
+    public bool SetHP(float value)
     {
         HP = value;
         if(HP <= 0)
